Adapt approaches for any target task
