+++ approaches/__init__.py
@@ -1,5 +1,7 @@
-from approaches.abstract_approach import AbstractApproach
+from approaches.duplicates_detection import DuplicatesDetectionApproach
 from approaches.pretraining_tasks_test import PretrainingTasksTest
 from approaches.intersection import IntersectionApproach
 from approaches.simple import SimpleApproach
 from approaches.tfidf import TfIdfApproach
+from approaches.abstract_approach import AbstractApproach
+from approaches.assignment import AssignmentApproach
