+++ approaches/pretraining_tasks_test.py
@@ -16,7 +16,7 @@ class PretrainingTasksTest(AbstractApproach):
     """
 
     def __init__(self, approach: AbstractApproach, pretraining_tasks: List[str]):
-        super(PretrainingTasksTest, self).__init__(approach.train, approach.test)
+        super(PretrainingTasksTest, self).__init__(approach.train, approach.test, approach.metrics)
         self.approach = approach
         self.pretraining_tasks = pretraining_tasks
 
@@ -43,11 +43,5 @@ class PretrainingTasksTest(AbstractApproach):
     def evaluate(self, model: AbstractModel, topks: List[int]) -> Dict[str, np.ndarray]:
         return self.approach.evaluate(model, topks)
 
-    def setup_approach(self):
-        pass
-
-    def get_duplicated_ids(self, query_num: int, topn: int) -> np.ndarray:
-        pass
-
-    def update_history(self, query_num: int):
-        pass
+    def save_results(self, save_to_path: str, model_name: str, plot: bool = False):
+        self.approach.save_results(save_to_path, model_name, plot)
